Use partition up as ready criterion for TransientGCP
wait_for_cluster_ready() was redefined in base class to wait for any node
be ready, but this will indefinitely hang for the TransientGCP backend, which
starts all nodes in powersave mode, and instead explicitly marks the partition
as up when it's ready.

Since TransientGCP's criterion is probably less common than actually waiting for
a node, let's keep the base method as-is and override it here.