Don't symlink absolute paths to workspace
If a file specified as an absolute path resides on the same NFS as the workspace,
we shouldn't symlink it into the workspace -- chances are the user meant to
supply it as an absolute path.

If the user truly wants to symlink, they can specify this manually with an override.