Don't assume job_id corresponds to Slurm task array ID
This is in the case of job avoidance -- if job_id 0 gets avoided, but job_id 1
gets re-run, then Slurm task array ID 0 -> job_id 1

We do, however, assume that ordering is conserved between the two. This is due
to the fact that Python dictionaries are guaranteed to be insertion ordered.