Remove common inputs dir. if job avoidance occurs
Otherwise, the localizer will try (and fail) to save common inputs to the
common directory. If job avoidance occurs, this directory will already be
present and thus crash the localizer.

Removing it wholesale is somewhat hacky, but it's a lot more straightforward
than modifying the orchestrator to coordinate this with the localizer.