Make sure config directory is empty before starting Slurm
We use sinfo to query whether Slurm is ready, and interpret a specific
failure message ("slurm_load_partitions: Unable to contact slurm controller
(connect failure)") as a sign that the controller node exists but has not yet
started slurmctld.

However, if slurm.conf is already present on the NFS disk and it points to
a nonexistent controller address, sinfo will crash with a different message.

Rather than trying to handle a menagerie of error messages (or simply treating
sinfo exiting nonzero as a sign, which is too heavyhanded), we simply delete
the Slurm config files before starting the Docker.