load_acct_from_disk() accounts for job avoidance
When loading .sacct from disk, we check to see if the shard was avoided (based
on .*exit_code, and self.job_spec[job] = None). If it was avoided, we override
whatever status got loaded from disk to "COMPLETED", since wolF checks to see
if all statuses == "COMPLETED" to infer whether the whole task finished
successfully.