Make same filesystem check more robust
Rather than distinguishing between different filesystems based on the device,
we distinguish based on the mountpoint.

We also bind mount /mnt/nfs so that it has a different mountpoint from /

Since only /mnt/nfs is exported, we cannot symlink files outside of /mnt/nfs
(but residing on the same device).