Job avoidance for scratch disks working for real
We needed to be a bit more clever about this, since the scratch disk still
needs to get mounted in order for outputs directory to be populated with the
correct symlinks.

Thus, we append `exit 15` to the standard localization script (which includes
scratch disk mounting). We also update the entrypoint to continue with delocalization
after detecting an `exit 15` from localization.sh