Revert "Continue with job when another instance finished localization"
This reverts commit a2eed8f15318bab6f1ba6f26e418da821a0c957c.
Suppose there are multiple tasks attempting to create the exact same disk. This
is usually only the case if multiple workflows are being run in parallel that
are trying to localize the same files via LocalizeToDisk. Only
one task will actually succeed at doing this; the other tasks will be waiting on
it.  Since LocalizeToDisk not actually run any script, we exit 15, which causes
entrypoint.sh to skip straight to the end.

Now suppose these multiple tasks actually do something in their script. In that
case, we wouldn't want to skip the script via exit 15. The problem with the change
proposed here is that it would skip necessary steps like mounting the disk.

We should address this issue later, but for now I'm rolling back this commit.