Skip hashing files that already exist on finished scratch disk
Job avoidance for finished scratch disks still requires mounting the disk,
since it's possible that the output directory on the NFS does not exist for
the task using the disk. This in turn requires running the delocalization script.

File on scratch disks can be quite large, so it can take a long time for
the delocalizer to hash them. We thus skip hashing if the delocalizer is running
on a finished scratch disk, and the hash .crc32 file already exists.

Previously, we skipped running the delocalizer altogether if we were running
on a finished scratch disk and the outputs directory existed. However, this
created an edge case where the outputs directory existed but nothing was
actually delocalized into it.